# Improve performace with `scripts/vim-patch.sh -l`