Differences in output:

- keeps/adds the "v" prefix, which allows to easily copy'n'paste this for "git show" in the Vim repo
- removes commit hashes for runtime file updates, e.g. 7f2e9d7c9cdfc5201a899b7b610edf64bf80c45f` (vim/vim@7f2e9d7c9cdfc5201a899b7b610edf64bf80c45f)

Before:
> 134.99s user 59.14s system 153% cpu 2:06.11 total

After:
> 20.88s user 33.04s system 106% cpu 50.469 total