@blueyed 2018-09-25
Updated.

It will now append ("+runtime") for hash-only commits, and removes "8.0.0649", which was reported before, but included in 8b543d09d.

Diff (using "diff /tmp/vim-patch.orig =(sed 's/• v/• /' /tmp/vim-patch-new)", i.e. compensating for the "v" prefix being added):

```
67d66
<   • 8.0.0649 (+runtime)
481c480
<   • 7f2e9d7c9cdfc5201a899b7b610edf64bf80c45f
---
>   • 7f2e9d7c9cdfc5201a899b7b610edf64bf80c45f (+runtime)
512c511
<   • a2a80162deb1e96e16b097dfe48b61b6eb0824bf
---
>   • a2a80162deb1e96e16b097dfe48b61b6eb0824bf (+runtime)
538c537
<   • 2f0584910c0ea93966dfa7b3b0272dfb16ea6f91
---
>   • 2f0584910c0ea93966dfa7b3b0272dfb16ea6f91 (+runtime)
556c555
<   • 4697399e8c805325009351a2488e3da530d0af38
---
>   • 4697399e8c805325009351a2488e3da530d0af38 (+runtime)
559c558
<   • f0b03c4e98f8a7184d8b4a5d702cbcd602426923
---
>   • f0b03c4e98f8a7184d8b4a5d702cbcd602426923 (+runtime)
571c570
<   • df980db69b831381c37c3e2973a6eefa10df305c
---
>   • df980db69b831381c37c3e2973a6eefa10df305c (+runtime)
578c577
<   • 40962ec9c0e7b8699e101182b06ddd39dc0e1212
---
>   • 40962ec9c0e7b8699e101182b06ddd39dc0e1212 (+runtime)
612c611
<   • 7254067ee970686cc3ff4a1effc3e49e9192a5c1
---
>   • 7254067ee970686cc3ff4a1effc3e49e9192a5c1 (+runtime)
655c654
<   • 22f1d0e35eb0e503c2dc7813a973c7e6ca1b7dde
---
>   • 22f1d0e35eb0e503c2dc7813a973c7e6ca1b7dde (+runtime)
659c658
<   • 3ad8772ef02e4352b0fa4e2697abc357734b62e1
---
>   • 3ad8772ef02e4352b0fa4e2697abc357734b62e1 (+runtime)
696c695
<   • b5b7562475ad032a174b893286172de0d2c157cd
---
>   • b5b7562475ad032a174b893286172de0d2c157cd (+runtime)
716c715
<   • 98ef233e14faaaf42e0342da1803d493ce0337b2
---
>   • 98ef233e14faaaf42e0342da1803d493ce0337b2 (+runtime)
753c752
<   • ab943431d8fcd856008a025b0e5652dd4b8007fc
---
>   • ab943431d8fcd856008a025b0e5652dd4b8007fc (+runtime)
755c754
<   • 25a6e8a769aa1c0d308b5f871132961f37986d0a
---
>   • 25a6e8a769aa1c0d308b5f871132961f37986d0a (+runtime)
767c766
<   • 51ad4eaa22e15cf1fe6c45d82c7e1371e00401a4
---
>   • 51ad4eaa22e15cf1fe6c45d82c7e1371e00401a4 (+runtime)
838c837
<   • 7dda86f2ff35bb80afce4da24782fd58216bbe50
---
>   • 7dda86f2ff35bb80afce4da24782fd58216bbe50 (+runtime)
878c877
<   • 15142e27aaafa15b72d1042c25fbb5e4f12b6736
---
>   • 15142e27aaafa15b72d1042c25fbb5e4f12b6736 (+runtime)
893c892
<   • 85eee130f44a2201d88ca2aeff0af3b11dd75fa9
---
>   • 85eee130f44a2201d88ca2aeff0af3b11dd75fa9 (+runtime)
913c912
<   • 7db25fed5de1be922b8cbb0328149469606a0424
---
>   • 7db25fed5de1be922b8cbb0328149469606a0424 (+runtime)
922c921
<   • eb3dc87f01391bb075d97aef3d00f91b4e08a25c
---
>   • eb3dc87f01391bb075d97aef3d00f91b4e08a25c (+runtime)
932c931
<   • 7c63fbc46e218cad266641a743a6c1353c2f00d4
---
>   • 7c63fbc46e218cad266641a743a6c1353c2f00d4 (+runtime)
936c935
<   • 2f01889055f14842c1d4f9abd91139fc6bd95f15
---
>   • 2f01889055f14842c1d4f9abd91139fc6bd95f15 (+runtime)
948c947
<   • 0b0f0992d46ef02fdfc1240744bef91d4299c1df
---
>   • 0b0f0992d46ef02fdfc1240744bef91d4299c1df (+runtime)
999c998
<   • d2f3a8b8787333abf2300d38836b196955f10c00
---
>   • d2f3a8b8787333abf2300d38836b196955f10c00 (+runtime)
1027c1026
<   • 675e8d6adb88d5b38087870dfda51c0205070741
---
>   • 675e8d6adb88d5b38087870dfda51c0205070741 (+runtime)
1056c1055
<   • 6dc819b1299e1d9f99303568772ade544d5c1322
---
>   • 6dc819b1299e1d9f99303568772ade544d5c1322 (+runtime)
1078c1077
<   • fd35811ca528de7612f9571bce20ef4c392fe5f7
---
>   • fd35811ca528de7612f9571bce20ef4c392fe5f7 (+runtime)
1098c1097
<   • b477af2260d9bc7ae7f743f0a14265d7ee12ee09
---
>   • b477af2260d9bc7ae7f743f0a14265d7ee12ee09 (+runtime)
1111c1110
<   • a9604e61451707b38fdcb088fbfaeea2b922fef6
---
>   • a9604e61451707b38fdcb088fbfaeea2b922fef6 (+runtime)
1133c1132
<   • 91f84f6e11cd879d43d651c0903d85bff95f0716
---
>   • 91f84f6e11cd879d43d651c0903d85bff95f0716 (+runtime)
1139c1138
<   • d2855f5454c5c6c5f786b228c5b67757edfefcb1
---
>   • d2855f5454c5c6c5f786b228c5b67757edfefcb1 (+runtime)
1175c1174
<   • d473c8c101262702ea9eeb14907ee20a786942b2
---
>   • d473c8c101262702ea9eeb14907ee20a786942b2 (+runtime)
1229c1228
<   • fc65cabb15d0236bce001ad78e12a40511caf941
---
>   • fc65cabb15d0236bce001ad78e12a40511caf941 (+runtime)
1240c1239
<   • 20aac6c1126988339611576d425965a25a777658
---
>   • 20aac6c1126988339611576d425965a25a777658 (+runtime)
1253c1252
<   • 93a1df2c205c8399d96c172d9483e0793d32892a
---
>   • 93a1df2c205c8399d96c172d9483e0793d32892a (+runtime)
```