@justinmk 2018-09-27
> removes "8.0.0649", which was reported before, but included in 8b543d0.

Ah, nice. You mean the old code incorrectly reported 8.0.0649 as missing?

> compensating for the "v" prefix being added):

Why add the v prefix? Would prefer not to, otherwise it's less convenient to copy/paste. AFAIR `vim-patch.sh -p` does not allow v-prefixed patch number.