@blueyed 2018-09-27
> Ah, nice. You mean the old code incorrectly reported 8.0.0649 as missing?

Yes.

> Why add the v prefix?

For copy'n'pasting (it is a valid ref in Vim sources), but `vim-patch.sh -p` does not accept it indeed (thought otherwise).  I can make it handle it, but it is also OK to not add the v prefix for me.
Let me know.