@justinmk 2018-09-30
> I can make it handle it, but it is also OK to not add the v prefix for me. Let me know.

Would prefer to leave it as it was, in case `vimpatch.lua` or other functions depend on the old output.