@justinmk 2018-10-28
Seems like the original performance improvement was because of the [removal](https://github.com/neovim/neovim/pull/9044/files#r219769902) of checking for non-tagged commits. Now it takes 1.5 minutes, just like the old script.

Still in favor of merging this since it improves the code a bit (but see the last couple of comments). Though using arrays in shell script is a red flag ... :)