@blueyed 2018-12-21
It is still faster for me:

new: scripts/vim-patch.sh -l > vim-patch-output.new  21.26s user 34.81s system 103% cpu 54.145 total
old: scripts/vim-patch.sh -l  138.57s user 73.01s system 160% cpu 2:11.70 total

`diff vim-patch-output.old =(sed 's/• v/• /' vim-patch-output.new)`:
```
954c954
<   • ba3ff539303c7bb6e46a6802dce3c7b2e55284e0
---
>   • ba3ff539303c7bb6e46a6802dce3c7b2e55284e0 (+runtime)
966c966
<   • f0d58efc9dc46be37c629cbc99b4125448ca39fd
---
>   • f0d58efc9dc46be37c629cbc99b4125448ca39fd (+runtime)
981c981
<   • b730f0c7ba36492d795f081b19bbcb85cdf0f50f
---
>   • b730f0c7ba36492d795f081b19bbcb85cdf0f50f (+runtime)
1004c1004
<   • d47d52232bf21036c5c89081458be7eaf2630d24
---
>   • d47d52232bf21036c5c89081458be7eaf2630d24 (+runtime)
1033c1033
<   • 9d87a37ee9d87f5bdbc779bc940d5f1e6f055d0a
---
>   • 9d87a37ee9d87f5bdbc779bc940d5f1e6f055d0a (+runtime)
```